+++ src/ui.cpp
@@ -4,7 +4,7 @@
 
 void render_bar(spritebatch& batch, bmfont& font, quad_renderer& quad, const rectangle& dest, std::string_view text, uint32_t bar_color, double value)
 {
-    //int x = 8, y = 8;
+    // int x = 8, y = 8;
 
     int x = dest.x;
     int y = dest.y;
@@ -17,7 +17,7 @@ void render_bar(spritebatch& batch, bmfont& font, quad_renderer& quad, const rec
     quad.draw_quad(dest, 0x140c1cff);
 
     // 2px margin either end
-    const int bar_width = value * (w - 2);
+    const int bar_width = (int)(value * (w - 2));
 
     quad.draw_quad(rectangle{x + 1, y + 1, bar_width, h - 2}, bar_color);
     quad.end();
